Move all unit tests from tests to subdirectory tests/unit
This will make it easier to add tests other than unit tests, e.g. typing tests, while keeping them separate.
