+++ tests/unit/exceptions/base_exceptions_test.py
