+++ tests/unit/exceptions/common_exceptions_test.py
