+++ tests/unit/exceptions/list_exceptions_test.py
